[MH-253] Refactors verified helper methods into AD model.
Also implements a number of factory methods to help with testing, and
refactors a couple tests to make use of them.

Also also moves the cloudfactory model into documents to avoid circular
references - these models are tightly bound as they are.
