+++ myhpom/admin.py
@@ -1,13 +1,12 @@
 from __future__ import unicode_literals
+
 from django.contrib import admin
 from django.core.urlresolvers import reverse
-from myhpom.models.state import State
-from myhpom.models.state_requirement import StateRequirement
-from myhpom.models.state_requirement_link import StateRequirementLink
-from myhpom.models.document import AdvanceDirective, DocumentUrl
-from myhpom.models.cloudfactory import CloudFactoryDocumentRun
-
 from django.utils.safestring import mark_safe
+
+from myhpom.models import (
+    AdvanceDirective, CloudFactoryDocumentRun, DocumentUrl, State, StateRequirement,
+    StateRequirementLink)
 
 
 class StateRequirementLinkAdmin(admin.TabularInline):
