+++ myhpom/models/__init__.py
@@ -1,7 +1,7 @@
-from .document import AdvanceDirective, DocumentUrl
+# flake8: noqa
+from .document import AdvanceDirective, DocumentUrl, CloudFactoryDocumentRun
 from .state import State
 from .user_details import UserDetails
 from .health_network import HealthNetwork
 from .state_requirement import StateRequirement
 from .state_requirement_link import StateRequirementLink
-from .cloudfactory import CloudFactoryDocumentRun
