+++ myhpom/tests/test_cloudfactory_models.py
@@ -1,4 +1,3 @@
-import json
 import os
 from django.test import TestCase
 from django.core.files.uploadedfile import SimpleUploadedFile
@@ -80,31 +79,3 @@ class DocumentRunModelTestCase(TestCase):
         run = CloudFactoryDocumentRun.objects.create(document_url=self.document_url)
         response_content = 'This content is not json, so not parsable'
         self.assertRaises(ValueError, run.save_response_content, response_content)
-
-    def test_is_successful_run(self):
-        SUCCESS_DATA = open(os.path.join(
-            os.path.dirname(__file__), 'fixtures/cloudfactory/callback_success.json')).read()
-
-        # If the status is equal to a successful run, and all the outputs are
-        # successful, then is_successful_run will return True
-        run = CloudFactoryDocumentRun.objects.create(
-            document_url=self.document_url)
-        run.save_response_content(SUCCESS_DATA)
-        self.assertTrue(run.is_successful_run)
-
-        # Even if the successful run is true, if the status is failed, so is
-        # this the run:
-        run.status = CloudFactoryDocumentRun.STATUS_ABORTED
-        run.save()
-        self.assertFalse(run.is_successful_run)
-
-        # If a one of the outputs are false, then the run is not successful.
-        failed_run = json.loads(SUCCESS_DATA)
-        failed_run['units'][0]['output']['owner_name_matches'] = False
-        run.save_response_content(json.dumps(failed_run))
-        self.assertFalse(run.is_successful_run)
-
-        # Especially if the status is aborted
-        failed_run['status'] = CloudFactoryDocumentRun.STATUS_ABORTED
-        run.save_response_content(json.dumps(failed_run))
-        self.assertFalse(run.is_successful_run)
